Environment class, moved camera, new material...
... some further implementation for RenderComponents, fixed
getComponent() in Entity, added getter for render context width and
height.

+++ DBGL_Base/src/DBGL/Window/Window.cpp
@@ -104,7 +104,7 @@ namespace dbgl
 	if (m_pRenderContext != NULL)
 	    delete m_pRenderContext;
 
-	m_pRenderContext = new RenderContext { getFrameWidth(), getFrameHeight() };
+	m_pRenderContext = new RenderContext { (unsigned int) getFrameWidth(), (unsigned int) getFrameHeight() };
     }
 
     void Window::init(Bitmask<> options)
